- updated weapon system file names - some code style modifications in a few other files - removed some compiler warnings

+++ code/branches/FICN/src/orxonox/core/ArgReader.cc
@@ -82,7 +82,7 @@ namespace orxonox {
     int argpos = checkOption(option) + 1;
     if(argpos != 0)
     {
-      floating = atof(arguments_[argpos]);
+      floating = (float)atof(arguments_[argpos]);
     }
     else
     {
